move ResourceType enum to wlm
