+++ server/src/main/java/org/opensearch/wlm/QueryGroupLevelResourceUsageView.java
@@ -8,7 +8,6 @@
 
 package org.opensearch.wlm;
 
-import org.opensearch.search.ResourceType;
 import org.opensearch.tasks.Task;
 
 import java.util.List;
