+++ server/src/test/java/org/opensearch/cluster/metadata/QueryGroupMetadataTests.java
@@ -14,8 +14,8 @@ import org.opensearch.core.common.io.stream.NamedWriteableRegistry;
 import org.opensearch.core.common.io.stream.Writeable;
 import org.opensearch.core.xcontent.XContentBuilder;
 import org.opensearch.core.xcontent.XContentParser;
-import org.opensearch.search.ResourceType;
 import org.opensearch.test.AbstractDiffableSerializationTestCase;
+import org.opensearch.wlm.ResourceType;
 
 import java.io.IOException;
 import java.util.Collections;
