+++ server/src/test/java/org/opensearch/wlm/ResourceTypeTests.java
@@ -6,14 +6,13 @@
  * compatible open source license.
  */
 
-package org.opensearch.search;
+package org.opensearch.wlm;
 
 import org.opensearch.action.search.SearchShardTask;
 import org.opensearch.core.tasks.resourcetracker.ResourceStats;
 import org.opensearch.tasks.CancellableTask;
 import org.opensearch.test.OpenSearchTestCase;
 
-import static org.junit.Assert.assertThrows;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
